Move parser definitions back to their original file to minimize diff.
